requirements.txt
# -----------------------------
# Core scientific stack
# -----------------------------
numpy>=1.23
scipy>=1.9
pandas>=1.5

# -----------------------------
# Deep Learning (PyTorch)
# -----------------------------
torch>=2.0
torchvision>=0.15

# -----------------------------
# Machine Learning (Traditional)
# -----------------------------
scikit-learn>=1.2
xgboost>=1.7
joblib>=1.2

# -----------------------------
# Image processing
# -----------------------------
Pillow>=9.4
opencv-python>=4.7

# -----------------------------
# Visualization
# -----------------------------
matplotlib>=3.6
seaborn>=0.12

# -----------------------------
# Utilities
# -----------------------------
tqdm>=4.64
